pwstream: use size_t for candidate indexes
Store password range indexes with the same type used for pool and mask lengths, and carry that type through brute-force iteration and candidate lookup. Preserve the existing split layout with overflow-safe quotient and remainder arithmetic.

Add coverage for ranges spanning the full size_t domain and document the split calculation with a worked example.

+++ lib/pwstream.h
@@ -21,8 +21,7 @@
 
 struct pwstream;
 struct entry {
-	/* TODO: convert to ssize_t */
-	int start, stop, initial;
+	size_t start, stop, initial;
 };
 
 int pwstream_new(struct pwstream **pws);
@@ -50,6 +49,6 @@ size_t pwstream_get_pwlen(const struct pwstream *pws);
 
 size_t pwstream_get_stream_count(const struct pwstream *pws);
 
-bool pwstream_is_empty(const struct pwstream *pws, unsigned int stream);
+bool pwstream_is_empty(const struct pwstream *pws, size_t stream);
 
 #endif /* _PWSTREAM_H_ */
